requirements.txt
Django==4.0.4
python-dateutil==2.8.2
rule-engine==3.4.0
six==1.16.0
djangorestframework==3.13.1
